requirements.txt
PyQt5>=5.15.0
requests>=2.28.0
Pillow>=9.0.0
piexif>=1.1.0
